Bodies

Methods

ViscousFlow.Bodies.Ellipse — Type

Ellipse(a,b,n) <: Body

Construct an elliptical body with semi-major axis a and semi-minor axis b, with n points distributed on the body perimeter.

ViscousFlow.Bodies.NACA4 — Type

NACA4(cam,pos,thick[;np=20][,Xc=(0.0,0.0)][,len=1.0]) <: Body{N}

Generates points in the shape of a NACA 4-digit airfoil of chord length 1. The relative camber is specified by cam, the position of maximum camber (as fraction of chord) by pos, and the relative thickness by thick.

The optional parameter np specifies the number of points on the upper or lower surface. The optional parameter Zc specifies the mean position of the vertices (which is set to the origin by default). The optional parameter len specifies the chord length.

Example

julia> w = Bodies.NACA4(0.0,0.0,0.12);

ViscousFlow.Bodies.Plate — Type

Plate(length,thick,n,[λ=1.0]) <: Body

Construct a flat plate with length length and thickness thick, with n points distributed on the body perimeter.

The optional parameter λ distributes the points differently. Values between 0.0 and 1.0 are accepted.

The constructor Plate(length,n,[λ=1.0]) creates a plate of zero thickness.

ViscousFlow.Bodies.RigidTransform — Type

RigidTransform(x::Tuple{Float64,Float64},α::Float64)

Construct a rigid-body transform operator, with rotation by angle α and translation specified by x. The translation coordinates are specified in the target coordinate system.

The resulting transform can be used as an operator on pairs of coordinate vectors, x and y, or on bodies. For transformation of bodies, it only overwrites the x and y fields of the body, but leaves the x̃ and ỹ (body coordinates) intact.

The translation can be provided as either a tuple (x,y) or as a complex number.

Constructors

RigidTransform((x,y),α)

RigidTransform(u::Vector{Float64})

RigidTransform(u::NTuple{3,Float64})

RigidTransform.(u) where u is a collection of vectors or tuples.

Example

julia> body = Bodies.Ellipse(0.5,0.1,100)
Elliptical body with 100 points and semi-axes (0.5,0.1)
   Current position: (0.0,0.0)
   Current angle (rad): 0.0

julia> T = RigidTransform((1.0,1.0),π/4)
Rigid-body transform
  Translation: (1.0,1.0)
  Rotation angle (rad): 0.7853981633974483

julia> T(body)
Elliptical body with 100 points and semi-axes (0.5,0.1)
   Current position: (1.0,1.0)
   Current angle (rad): 0.7853981633974483

ViscousFlow.Bodies.Circle — Method

Circle(a,n) <: Body

Construct a circular body with radius a and with n points distributed on the body perimeter.

ViscousFlow.Bodies.centraldiff — Method

centraldiff(body::Body/BodyList) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the circular central differences of coordinates on body body (or on each body in list body).

ViscousFlow.Bodies.dlength — Method

dlength(body::Body/BodyList) -> Vector{Float64}

Compute the lengths of the faces on the perimeter of body body, whose ends are at the current x and y coordinates (in inertial space) of the body. Face 1 corresponds to the face between points 1 and 2, for example.

ViscousFlow.Bodies.dlengthmid — Method

dlengthmid(body::Body/BodyList) -> Vector{Float64}

Compute the lengths of the faces formed between the face midpoints on the perimeter of body body. The indexing of these midpoint faces is consistent with that of the regular vertex points adjacent to both midpoints. Midpoint face 2 corresponds to the face between midpoints 1 and 2, for example.

ViscousFlow.Bodies.getrange — Method

getrange(bl::BodyList,i::Int) -> Range

Return the range of indices in the global set of Lagrange point data corresponding to body i in body list bl.

ViscousFlow.Bodies.midpoints — Method

midpoints(body::Body/BodyList) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the x and y midpoints of the faces on the perimeter of body body, whose ends are at the current x and y coordinates (in inertial space) of the body. Face 1 corresponds to the face between points 1 and 2, for example.

If body is a BodyList, then it computes the differences separately on each constituent body.

ViscousFlow.Bodies.normal — Method

normal(body::Body/BodyList) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the current normals (in inertial components) of the faces on the perimeter of body body, whose ends are at the current x and y coordinates (in inertial space) of the body. Face 1 corresponds to the face between points 1 and 2, for example.

ViscousFlow.Bodies.normalmid — Method

normalmid(body::Body/BodyList) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the current normals (in inertial components) of the faces formed between midpoints on the perimeter of body body (or each body in list body).

Base.collect — Method

collect(bl::bodylist) -> Vector{Float64}, Vector{Float64}

Collect the inertial-space coordinates of all of the Lagrange points comprising the bodies in body list bl and return each assembled set of coordinates as a vector.

Base.diff — Method

diff(body::Body/BodyList) -> Tuple{Vector{Float64},Vector{Float64}}

Compute the x and y differences of the faces on the perimeter of body body, whose ends are at the current x and y coordinates (in inertial space) of the body. Face 1 corresponds to the face between points 1 and 2, for example.

If body is a BodyList, then it computes the differences separately on each constituent body.

Base.length — Method

length(body::Body)

Return the number of points on the body perimeter

Base.sum — Method

sum(f::AbstractVector,bl::BodyList,i::Int) -> Real

Compute a sum of the elements of vector f corresponding to body i in body list bl.

Base.vec — Method

vec(T::RigidTransform) -> Vector{Float64}

Returns a length-3 vector of the form [x,y,α] corresponding to the translation and rotation specified by the given transform T.

Base.vec — Method

vec(tl::Vector{RigidTransform}) -> NTuple{N,Vector{Float64}}

Returns a tuple of length-3 vectors of the form [x,y,α] corresponding to the translation and rotation specified by the given by the list of transforms tl.

Base.view — Method

view(f::AbstractVector,bl::BodyList,i::Int) -> SubArray

Provide a view of the range of values in vector f corresponding to the Lagrange points of the body with index i in body list bl.

Index

ViscousFlow.Bodies.Ellipse

ViscousFlow.Bodies.NACA4

ViscousFlow.Bodies.Plate

ViscousFlow.Bodies.RigidTransform

Base.collect

Base.diff

Base.length

Base.sum

Base.vec

Base.vec

Base.view

ViscousFlow.Bodies.Circle

ViscousFlow.Bodies.centraldiff

ViscousFlow.Bodies.dlength

ViscousFlow.Bodies.dlengthmid

ViscousFlow.Bodies.getrange

ViscousFlow.Bodies.midpoints

ViscousFlow.Bodies.normal

ViscousFlow.Bodies.normalmid