Functions and types

Functions

ViscousFlow.setup_grid — Method

setup_grid(xlim::Tuple,ylim::Tuple,phys_params::Dict)

Construct a Cartesian grid with limits xlim and ylim and spacing determined by the Reynolds number in the phys_params.

ViscousFlow.surface_point_spacing — Method

surface_point_spacing(g::PhysicalGrid,phys_params)

Calculate the surface point spacing for a given grid, using the specified parameter "point spacing ratio" in the physical parameters, or the default value (1.4) if not specified.

ViscousFlow.viscousflow_system — Method

viscousflow_system(grid,[bodies];kwargs...)

Construct the operators and cache variables for a viscous flow problem.

The kwargs are the optional keyword aguments. There are several, some of which are crucial for certain types of problems.

ddftype = to set the DDF type. The default is CartesianGrids.Yang3.

scaling = to set the scaling type, IndexScaling (default) or GridScaling.

dtype = to set the element type to Float64 (default) or ComplexF64.

phys_params = A dictionary to pass in physical parameters

bc = A dictionary to pass in boundary condition data or functions, using "external" and "internal" keys to pass in functions that provide the corresponding surface data outside and inside the surface(s).

forcing = A dictionary to pass in forcing models (via the "forcing models" key), or to pass in an alternative model for the freestream velocity (with the "freestream" key)

motions = to provide function(s) that specify the velocity of the immersed surface(s). Note: if this keyword is used, it is assumed that surfaces will move.

timestep_func = to pass in a function for time-dependent problems that provides the time-step size. It is expected that this function takes in two arguments, the grid::PhysicalGrid and phys_params, and returns the time step. It defaults to the basic Fourier/CFL type function default_timestep

Types

ViscousFlow.ViscousIncompressibleFlowProblem — Type

ViscousIncompressibleFlowProblem

ILM problem type dealing with vector-type data.

Index

ViscousFlow.ViscousIncompressibleFlowProblem

ViscousFlow.setup_grid

ViscousFlow.surface_point_spacing

ViscousFlow.viscousflow_system