Functions and types

Functions

Base.size — Method

size(sys::NavierStokes,d::Int) -> Int

Return the number of indices of the grid used by sys along dimension d.

Base.size — Method

size(sys::NavierStokes) -> Tuple{Int,Int}

Return a tuple of the number of indices of the grid used by sys

CartesianGrids.cellsize — Method

cellsize(sys::NavierStokes) -> Float64

Return the grid cell size of system sys

CartesianGrids.origin — Method

origin(sys::NavierStokes) -> Tuple{Int,Int}

Return a tuple of the indices of the primal node that corresponds to the physical origin of the coordinate system used by sys. Note that these indices need not lie inside the range of indices occupied by the grid. For example, if the range of physical coordinates occupied by the grid is (1.0,3.0) x (2.0,4.0), then the origin is not inside the grid.

RigidBodyTools.assign_velocity! — Method

assign_velocity!(u,v,x,y,m::DirectlySpecifiedMotion,t)

Given a DirectlySpecifiedMotion, vectors of x and y coordinates, and time t, return the velocities u and v at those points.

ViscousFlow.convective_derivative — Function

convective_derivative(integrator)

Return the convective derivative associated with integrator at its current state.

ViscousFlow.convective_derivative — Method

convective_derivative(sol::ODESolution,sys::NavierStokes,t)

Return the convective derivative associated with solution vector sol on the grid in sys, at time(s) t.

ViscousFlow.flowside — Method

flowside(sys::NavierStokes) -> FlowSide

Returns the side of the body on which the flow is to be computed.

ViscousFlow.force — Method

force(sol,sys::NavierStokes,bodyindex) -> Tuple{Vector,Vector}

Given the solution history vector sol and the system sys, return the force history on the body with index bodyindex as a tuple of vectors, one for each component.

ViscousFlow.force — Method

force(integ,bodyindex)

Given the state of the system in integrator integ, return the current force on the body with index bodyindex.

ViscousFlow.freestream — Method

freestream(t,sys::NavierStokes) -> Tuple

Return the value of the freestream in sys at time t as a tuple.

ViscousFlow.moment — Method

moment(sol,sys::NavierStokes,bodyindex[,center=(0,0)]) -> Vector

Given the solution history vector sol and the system sys, return the moment history on the body with index bodyindex as a vector. The center of the moment can be passed as an optional tuple argument.

ViscousFlow.moment — Method

moment(integ,bodyindex[,center=(0,0))

Given the state of the system in integrator integ, return the current moment on the body with index bodyindex. The center of the moment can be passed as an optional tuple argument.

ViscousFlow.newstate — Method

newstate(s::AbstractSpatialField,sys::NavierStokes)

Return an instance of the state vector for sys, assigned the data in the spatial field s.

ViscousFlow.newstate — Method

newstate(sys::NavierStokes)

Return a new (zero) instance of the state vector for sys.

ViscousFlow.pressure — Function

pressure(integrator)

Return the pressure field associated with integrator at its current state.

ViscousFlow.pressure — Method

pressure(sol::ODESolution,sys::NavierStokes,t)

Return the pressure field associated with solution vector sol on the grid in sys, at time(s) t.

ViscousFlow.pressurejump — Function

pressurejump(integrator)

Return the pressure jump at all of the surface points associated with integrator at its current state.

ViscousFlow.pressurejump — Method

pressurejump(sol::ODESolution,sys::NavierStokes,t)

Return the pressure jump at all of the surface points associated with solution sol on grid in sys at time(s) t.

ViscousFlow.scalarpotential — Function

scalarpotential(integrator)

Return the scalar potential field associated with integrator at its current state.

ViscousFlow.scalarpotential — Method

scalarpotential(sol::ODESolution,sys::NavierStokes,t)

Return the scalar potential field associated with solution vector sol on the grid in sys, at time(s) t.

ViscousFlow.setstepsizes — Method

setstepsizes(Re[,gridRe=2][,cfl=0.5][,fourier=0.5]) -> Float64, Float64

Set the grid cell spacing and time step size based on the Reynolds number Re, the grid Reynolds number gridRe, cfl number cfl, and grid Fourier number fourier. The last three parameters all have default values.

Example

Here is an example of setting parameters based on Reynolds number 100 (with default choices for grid Reynolds number, CFL number, and Fourier number):

julia> Δx, Δt = setstepsizes(100)
(0.02, 0.01)

ViscousFlow.streamfunction — Function

streamfunction(integrator)

Return the streamfunction field associated with integrator at its current state.

ViscousFlow.streamfunction — Method

streamfunction(sol::ODESolution,sys::NavierStokes,t)

Return the streamfunction field associated with solution vector sol on the grid in sys, at time(s) t.

ViscousFlow.timerange — Method

timerange(tf,sys::NavierStokes)

Create a range of times, starting at the t = Δt (the time step of sys), and ending at t = tf.

ViscousFlow.timestep — Method

timestep(sys::NavierStokes) -> Float64

Return the time step size of system sys

ViscousFlow.traction — Function

traction(integrator)

Return the traction at all of the surface points associated with integrator at its current state.

ViscousFlow.traction — Method

traction(sol::ODESolution,sys::NavierStokes,t)

Return the traction at all of the surface points associated with solution sol on grid in sys at time(s) t.

ViscousFlow.velocity — Function

velocity(integrator)

Return the velocity field associated with integrator at its current state.

ViscousFlow.velocity — Method

velocity(sol::ODESolution,sys::NavierStokes,t)

Return the velocity field associated with solution vector sol on the grid in sys, at time(s) t.

ViscousFlow.vorticity — Function

vorticity(integrator)

Return the vorticity field associated with integrator at its current state.

ViscousFlow.vorticity — Method

vorticity(sol::ODESolution,sys::NavierStokes,t)

Return the vorticity field associated with solution vector sol on the grid in sys, at time(s) t.

Types

ViscousFlow.NavierStokes — Type

mutable struct NavierStokes{NX, NY, N, MT<:ViscousFlow.PointMotionType, FS<:ViscousFlow.FreestreamType, SD<:ViscousFlow.FlowSide, DDF<:CartesianGrids.DDFType, FT, SP, WP, VP, SPP, GVP, VBP, SBP, FST, RHST, CDT, COT, DLT, VCT, PCT, SCT, SVCT, STCT}

A system type that utilizes a grid of NX x NY dual cells and N Lagrange forcing points to solve the discrete Navier-Stokes equations in vorticity form. The parameter static_points specifies whether the forcing points remain static in the grid. It should be set to false if a supplied motion requires that the points move.

Constructors:

NavierStokes(Re,Δx,xlimits,ylimits,Δt [,freestream = (0.0, 0.0)] [,pulses=nothing]) specifies the Reynolds number Re, the grid spacing Δx, the dimensions of the domain in the tuples xlimits and ylimits (excluding the ghost cells), and the time step size Δt. The other arguments are optional. The freestream argument can be passed as either a tuple (a static freestream) or a RigidBodyMotion for a time-varying freestream. The pulses argument can be used to pass in one or more spatiotemporal pulses.

NavierStokes(Re,Δx,xlimits,ylimits,Δt,bodies::Body/BodyList [,flow_side=ExternalInternalFlow] [,ddftype=CartesianGrids.Yang3]) passes the body information. This constructor sets the motions of the body/ies to be stationary. The same optional arguments used for the basic constructor also apply for this one. In addition, the flow_side can be set to ExternalFlow (default), InternalFlow, or ExternalInternalFlow. However, it is forced to ExternalInternalFlow for open Bodies (like Plate type).

NavierStokes(Re,Δx,xlimits,ylimits,Δt,bodies::Body/BodyList, motions::Motion/MotionList [,static_points=false]) passes the body and associated motion information. The list of motions must be the same length as the list of bodies. The same optional arguments used for the other constructors also apply for this one. In addition, static_points can be set to true if the supplied motion should not cause the points to move.

Index

ViscousFlow.NavierStokes

Base.size

Base.size

CartesianGrids.cellsize

CartesianGrids.origin

RigidBodyTools.assign_velocity!

ViscousFlow.convective_derivative

ViscousFlow.convective_derivative

ViscousFlow.flowside

ViscousFlow.force

ViscousFlow.force

ViscousFlow.freestream

ViscousFlow.moment

ViscousFlow.moment

ViscousFlow.newstate

ViscousFlow.newstate

ViscousFlow.pressure

ViscousFlow.pressure

ViscousFlow.pressurejump

ViscousFlow.pressurejump

ViscousFlow.scalarpotential

ViscousFlow.scalarpotential

ViscousFlow.setstepsizes

ViscousFlow.streamfunction

ViscousFlow.streamfunction

ViscousFlow.timerange

ViscousFlow.timestep

ViscousFlow.traction

ViscousFlow.traction

ViscousFlow.velocity

ViscousFlow.velocity

ViscousFlow.vorticity

ViscousFlow.vorticity