Functions and types

Functions

ViscousFlow.evaluate_freestream — Method

evaluate_freestream(t,phys_params)

Provide the components of the free stream velocity at time t. If the problem is set up in the rotational frame of reference, then this transforms the specified velocity and also subtracts the translational motion of the center of rotation.

ViscousFlow.extracted_power — Method

extracted_power(sol,sys,bodyi)

Calculated the power extracted from the flow (or power input into the flow, if negative) by body bodyi from the computational solution sol of system sys. This computes the power from the force and velocity due to prescribed oscillations. It does not include the power due to steady body motion.

ViscousFlow.force — Method

force(sol,sys,bodyi[;inertial=true]) -> Tuple{Vector}

Calculated the force exerted by the fluid on body bodyi from the computational solution sol of system sys. It returns the force history as a tuple of arrays: one array for each component. If inertial=true (default), then the components are provided in the inertial coordinate system. Otherwise, they are in the body coordinate system.

ViscousFlow.moment — Method

moment(sol,sys,bodyi[;center=(0,0)]) -> Vector

Calculated the momemt exerted by the fluid on body bodyi from the computational solution sol of system sys. It returns the moment history as an array. The moment is calculated about center center, which defaults to (0,0) in whichever coordinate system the problem is solved.

ViscousFlow.power — Method

power(sol,sys,bodyi)

Calculated the history of the total rate of work done by the flow on body bodyi (or on the flow by the body, if negative) from the computational solution sol of system sys.

ViscousFlow.setup_grid — Method

setup_grid(xlim::Tuple,ylim::Tuple,phys_params::Dict[;nthreads_max=length(Sys.cpu_info())])

Construct a Cartesian grid with limits xlim and ylim and spacing determined by the Reynolds number in the phys_params. The maximum number of threads can be optionally set; it defaults to the number of processor cores.

ViscousFlow.surface_point_spacing — Method

surface_point_spacing(g::PhysicalGrid,phys_params)

Calculate the surface point spacing for a given grid, using the specified parameter "point spacing ratio" in the physical parameters, or the default value (1.4) if not specified.

ViscousFlow.surface_velocity_in_translating_frame! — Method

surface_velocity_in_translating_frame!(vel,t,base_cache,phys_params)

Evaluates the surface velocity when the problem is set up in a moving frame of reference (specified with the reference frame keyword). It removes the translational velocity of the center of rotation, since this is applied as a free stream velocity (with change of sign).

ViscousFlow.transform_vector_to_inertial_coordinates — Method

transform_vector_to_rotating_coordinates(u,v,t,phys_params) -> Tuple

Transform a vector expressed in rotating coordinates with components u and v to the inertial coordinate system at time t.

ViscousFlow.transform_vector_to_rotating_coordinates — Method

transform_vector_to_rotating_coordinates(u,v,t,phys_params) -> Tuple

Transform a vector expressed in inertial coordinates with components u and v to the rotating coordinate system at time t.

ViscousFlow.velocity_rel_to_rotating_frame! — Method

velocity_rel_to_rotating_frame!(u_prime,t,sys::ILMSystem)

Changes the input velocity u_prime (u', which is measured relative to the translating frame) to û (measured relative to the translating/rotating frame)

ViscousFlow.viscousflow_system — Method

viscousflow_system(grid,[bodies];kwargs...)

Construct the operators and cache variables for a viscous flow problem.

The kwargs are the optional keyword aguments. There are several, some of which are crucial for certain types of problems.

ddftype = to set the DDF type. The default is CartesianGrids.Yang3.

scaling = to set the scaling type, GridScaling (default) or IndexScaling.

dtype = to set the element type to Float64 (default) or ComplexF64.

phys_params = A dictionary to pass in physical parameters, or to pass in alternative models for the freestream velocity (with the "freestream" key) or overall rotational motion (with the "rotation" key)

bc = A dictionary to pass in boundary condition data or functions, using "external" and "internal" keys to pass in functions that provide the corresponding surface data outside and inside the surface(s).

forcing = A dictionary to pass in forcing models (via the "forcing models" key)

motions = to provide function(s) that specify the velocity of the immersed surface(s). Note: if this keyword is used, it is assumed that surfaces will move.

timestep_func = to pass in a function for time-dependent problems that provides the time-step size. It is expected that this function takes in two arguments, the grid::PhysicalGrid and phys_params, and returns the time step. It defaults to the basic Fourier/CFL type function default_timestep

Types

ViscousFlow.ViscousIncompressibleFlowProblem — Type

ViscousIncompressibleFlowProblem

ILM problem type dealing with vector-type data.

Index

ViscousFlow.ViscousIncompressibleFlowProblem

ViscousFlow.evaluate_freestream

ViscousFlow.extracted_power

ViscousFlow.force

ViscousFlow.moment

ViscousFlow.power

ViscousFlow.setup_grid

ViscousFlow.surface_point_spacing

ViscousFlow.surface_velocity_in_translating_frame!

ViscousFlow.transform_vector_to_inertial_coordinates

ViscousFlow.transform_vector_to_rotating_coordinates

ViscousFlow.velocity_rel_to_rotating_frame!

ViscousFlow.viscousflow_system